Update Function.md
MEP 9/7/21 
Can the output from a function be a type (eg creditRisk),  a class (eg Account), a collection of classes (eg one or more ErrorMessage)? 
Can there be multiple outputs from a function? 
Can the return value be optional too?
So if everything is initially designed as a set of abstract functions one could (graphically?) see the dependencies between them even if there is no implementation? And potentially even discover any mismatches between the data types?